requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.6
django-embed-video==1.4.0
django-heroku==0.3.1
dnspython==2.1.0
email-validator==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.19.5
phonenumbers==8.12.19
Pillow==8.1.0
psycopg2==2.8.6
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
